move ApplicationsChoice to config.py
