+++ ykman/cli/config.py
@@ -27,8 +27,7 @@
 
 from __future__ import absolute_import
 
-from .util import (
-    click_postpone_execution, click_force_option, ApplicationsChoice)
+from .util import click_postpone_execution, click_force_option, UpperCaseChoice
 from ..device import device_config, FLAGS
 from ..util import APPLICATION
 from binascii import a2b_hex, b2a_hex
@@ -41,6 +40,15 @@ logger = logging.getLogger(__name__)
 
 
 CLEAR_LOCK_CODE = '0' * 32
+
+
+class ApplicationsChoice(UpperCaseChoice):
+    """
+    Special version of UpperCaseChoice that accepts openpgp as OPGP
+    """
+    def convert(self, value, param, ctx):
+        return 'OPGP' if value.lower() == 'openpgp' \
+            else super().convert(value, param, ctx)
 
 
 def prompt_lock_code(prompt='Enter your lock code'):
